+++ packages/uilib/src/lib/+layout.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
   import "$lib/style.css"
   import { page } from "$app/stores"
+  import { Theme } from "./theme"
 
   const modules = import.meta.glob("./(components|plugins)/**/+page.svelte")
   $: paths = Object.keys(modules)
@@ -35,6 +36,7 @@
   <title>OpenSCD Plugin Library (OPL)</title>
 </svelte:head>
 
+<Theme>
 <main>
   <sidebar>
     <nav>
@@ -53,6 +55,7 @@
     <slot />
   </content>
 </main>
+</Theme>
 
 <style>
   main {
@@ -60,6 +63,8 @@
     grid-template-columns: 200px 1fr;
     padding: 0;
     margin: 0;
+    height: 100vh;
+    overflow: hidden;
   }
 
   sidebar {
@@ -67,6 +72,7 @@
     border-right: thin gray solid;
     text-align: center;
     padding: 0.5rem 0;
+    overflow: auto;
   }
 
   sidebar ul {