+++ packages/uilib/src/lib/components/diagram/ied-element/+page.svelte
@@ -3,7 +3,14 @@
 	import { Example } from "../../internal"
 	import type { IEDNode } from "../nodes"
 
-	const node: IEDNode = {id: "0", width: 200, height: 100, label: "some longer IED name", isRelevant: true}
+	const node: IEDNode = {
+		id: "0", 
+		width: 200, 
+		height: 100, 
+		label: "some longer IED name", 
+		isRelevant: true,
+		children:[],
+	}
 
 	let containerWidth = 200
 	let containerHeight = 30