+++ packages/uilib/src/lib/components/internal/example/example.svelte
@@ -1,16 +1,16 @@
 <script>
-    import Theme from "../../../theme/theme.svelte"
+
     export let name = ""
 </script>
 
-<Theme>
-    <example>
-        <div>
-            <h3>{name}</h3>
-        </div>
-        <slot />
-    </example>
-</Theme>
+
+<example>
+    <div>
+        <h3>{name}</h3>
+    </div>
+    <slot />
+</example>
+
 
 <style>
     example {
@@ -18,6 +18,8 @@
         margin-bottom: 3rem;
         position: relative;
         width: 100%;
+        max-height: 100vh;
+        overflow: hidden;
     }
     h3 {
         margin-bottom: 1rem;