+++ packages/uilib/src/lib/plugins/communication-explorer/_func-layout-calculation/node-layout-ieds.ts
@@ -33,6 +33,7 @@ export function generateIEDLayout(
 			height:     config.height,
 			label:      ied.iedName,
 			isRelevant: isRelevant,
+			children:   [],
 		}
 	})