+++ packages/uilib/src/lib/plugins/network-explorer/store/store.ts
@@ -1,0 +1,15 @@
+import type { IEDNetworkInfoV3 } from "@oscd-plugins/core"
+import type { Edge, Node } from "@xyflow/svelte"
+import { writable } from "svelte/store"
+
+
+export class DiagramStore {
+	public nodes = writable<Node[]>([])
+	public edges = writable<Edge[]>([])
+	
+	public selectedNodes = writable<SelectedNode[]>([])
+}
+
+export type SelectedNode = IEDNetworkInfoV3 & {
+	connectedIEDs: string[]
+}